+++ tests/test_huggingface_jobs.py
@@ -113,7 +113,7 @@ def test_huggingface_job_config_can_select_best_gpu_profile(tmp_path):
     assert config.hardware_flavor == BEST_GPU_HARDWARE_FLAVOR
     assert config.dataset_preset == "first"
     assert config.device == "cuda"
-    assert config.precision == "float16"
+    assert config.precision == "bfloat16"
     assert config.batch_size == 64
     assert config.sampling_steps == 64
     assert config.max_train_steps == 5000
